+++ vnext/Desktop.ABITests/PerfTests.cpp
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#include "pch.h"
 #include <CppUnitTest.h>
 #include <winrt/facebook.react.h>