+++ vnext/Desktop.ABITests/pch.h
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#ifndef NOGDI
+#define NOGDI
+#endif
+#include <windows.h>